requirements.txt
anthropic
fastapi
httpx
uvicorn
PyYAML
websockets
